Initial stub commit
Very basic environment variable extraction and parsing.

The top-level `main.go` is a sample stub plugin to make ad-hoc testing easier.  It's not clear if this will be kept for "self-exampling" or not... but it seems useful for now.

The README will need updating (obviously).

Refactoring and tests are a must!